Parse ORE indexes as JSON arrays of hex-encoded strings
This change updates `cs_ore_64_8_v1` to parse ORE indexes (the
`'o'` field) as JSON arrays of hex-encoded strings (instead of casting
from the Postgres text format).

The corresponding change for encoding ORE indexes as JSON arrays
of hex-encoded stings has already been merged in Proxy.

This is similar to some of the changes in
https://github.com/cipherstash/encrypt-query-language/pull/86, but
we're parsing into the composite types for ORE indexes instead of
into a plain `bytea[]`. Parsing into the composite type allows for
ordering with an operator class on the output from `cs_ore_64_8_v1`.